исправление замечаний по ревью

+++ internal/gophermart/model/order.go
@@ -24,7 +24,7 @@ type Order struct {
 }
 
 type Accrual struct {
-	Order   string `json:"order"`
-	Status  Status `json:"status"`
-	Accrual *int   `json:"accrual,omitempty"`
+	Order   string   `json:"order"`
+	Status  Status   `json:"status"`
+	Accrual *float64 `json:"accrual,omitempty"`
 }
